requirements.txt
streamlit
transformers
torch
pillow
pandas
deep-translator
